Tweaking tsconfig.json, small formatting improvements
