+++ .gitignore
@@ -41,3 +41,4 @@ next-env.d.ts
 *.code-workspace
 
 demo/
+.idea/
