+++ src/app/privacy-policy/page.tsx
@@ -1,12 +1,15 @@
-import { getDocument } from "../../db/db";
+"use client"
+
+import {useContext} from "react";
 import Text from "@/components/text/text";
 import Article from "@/components/article/article";
-import type { Config, Text as TextType } from "@/lib/configProvider";
+import {Config, ConfigContext} from "@/lib/configProvider";
 
 import "./privacy_policy.css";
 
+
 export default async function PrivacyPolicy() {
-  const { privacy_policy }: Config | undefined = await getDocument<Config>("config", "config");
+  const { privacy_policy }: Config = useContext(ConfigContext)
 
   return (
     <Article className="container">
