+++ src/app/reviews/page.tsx
@@ -1,12 +1,34 @@
 import { Company } from "./columns";
-import { getDocument } from "../../db/db";
-import { ColumnKeys } from "./columns";
+import { getDocument } from "@/db/db";
 import DataTable from "@/components/aria-table/data-table";
 import Article from "@/components/article/article";
 
 export default async function Reviews() {
   const reviewData: Company | undefined = await getDocument<Company>("index", "properties_and_companies_index");
   const tableCaption = "Rental Reviews Data";
+
+    const ColumnKeys = [
+        {
+            key: "name",
+            title: "Name",
+        },
+        {
+            key: "company_type",
+            title: "Company Type",
+        },
+        {
+            key: "average_rating",
+            title: "Average Rating",
+        },
+        {
+            key: "adjusted_average_rating",
+            title: "Adjusted Average Rating",
+        },
+        {
+            key: "review_count",
+            title: "Review Count",
+        },
+    ];
 
   return (
     <Article className="flex flex-col container justify-center text-center container mx-auto py-10">
